Refactor crawler
