+++ indexer/Dockerfile
@@ -1,8 +1,8 @@
 FROM ruby:2.3
 
 RUN gem install pg nokogiri addressable work_queue redis i18n
-ADD app.rb /tmp/app.rb
+ADD . /tmp/.
 WORKDIR /tmp/
-RUN chmod 777 app.rb
+RUN chmod 777 main.rb
 
-CMD ruby app.rb
+CMD ruby main.rb
